fix lua bindings warnings

+++ cocos/base/CCString.h
@@ -167,7 +167,7 @@ public:
      *          it means that you needn't do a release operation unless you retain it.
      * @js NA
      */
-    static __String* createWithData(const unsigned char* pData, int nLen);
+    static __String* createWithData(const unsigned char* pData, size_t nLen);
 
     /** create a string with a file, 
      *  @return A String pointer which is an autorelease object pointer,
